rename resourceCollection to Resource reader and add an interface and split different class to different files
