+++ src/Microsoft.DocAsCode.Build.Engine/TemplateProcessors/Preprocessors/TemplatePreprocessorResource.cs
