+++ src/Microsoft.DocAsCode.Build.Engine/TemplateProcessors/TemplateResourceInfo.cs
