+++ src/Microsoft.DocAsCode.Build.Engine/TemplateProcessors/TemplateType.cs
