+++ src/Microsoft.DocAsCode.Build.Engine/TemplateProcessors/ViewRenderers/ITemplateRenderer.cs
