+++ src/Microsoft.DocAsCode.Build.Engine/TemplateProcessors/ViewRenderers/TemplateRendererResource.cs
