+++ src/Microsoft.DocAsCode.Build.Engine/TemplateProcessors/ViewRenderers/TemplateRendererType.cs
